backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
